runtimeVM: Fix shimv2 binary name construction
shim v2 expected the runtime name to be in the following format:
```
    ($dir.)?$prefix.$name.$version
    -------- ------ ----- ---------
        |     |     |      |
        v     |     |      |
"/usr/local/bin"    |      |
  (optional)  |     |      |
              v     |      |
       "containerd.shim."  |
                    |      |
                    v      |
               "kata-qemu" |
                           v
                          "v2"
```

The current code simply replaces all the "-" by ".", which ends up
working well as long as the $name doesn't have a "-" in it, which is, by
luck, what's done by the Kata Operator.  However, whenever a name has
"-" in it, which we can see happening on upstream kata-deploy - which
uses $name as kata-qemu, for instance - it'll simply make containerd
library confused, not being able to invoke the binary.

With everything said above in mind, let's introduce a specialized
function to build the shim v2 path.

Fixes: #4589

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>